move airporthighlights to components
